+++ app/src/main/java/com/copychrist/app/prayer/ui/contactgroups/ContactsPresenter.java
@@ -7,14 +7,14 @@ import com.copychrist.app.prayer.ui.BasePresenter;
  */
 
 public interface ContactsPresenter extends BasePresenter<ContactsView> {
-    void onContactGroupClicked(int contactGroupId);
+    void onContactGroupClicked(long contactGroupId);
     void onAddNewContactGroupClick();
     void onEditContactGroupClick();
-    void onSaveContactGroupClick(int contactGroupId, String groupName, String groupDesc, int sortOrder);
+    void onSaveContactGroupClick(long contactGroupId, String groupName, String groupDesc, int sortOrder);
     void onDeleteContactGroupClick();
     void onDeleteContactGroupConfirmed();
     void onAddNewContactClick();
     void onSaveContactClick(String firstName, String lastName, String groupName, String pictureUrl);
-    void onContactClick(int contactId);
-    void onPrayerRequestClick(int prayerRequestId);
+    void onContactClick(long contactId);
+    void onPrayerRequestClick(long prayerRequestId);
 }
